python/requirements.txt
# EYWA Client
eywa-client>=0.4.1

# Web Scraping
selenium>=4.0.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP Requests
requests>=2.31.0
aiohttp>=3.8.0

# Utilities
python-dateutil>=2.8.0
